[NUNCOTA-50] Add grub package to work with GRUB environment variables
This package gets/sets/unsets GRUB environment variables. To store any
changes Store function should be called. The GRUB environment file is
quite simple: it contains a string header, list of variables in
name=value format and all free space is filled with #. ESC symbol '\' is
handled in different way: in the file one '\' stores as '\\'.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>