[NUNCOTA-189] Implement get partition info in utils/partition package
Implement getting partition info (device, label, partuuid) in
utils/partition package. This info is required to get/set current boot
partition.

In order to perform unit tests, testtools partition is reimplemented in the
following way: instead of creating separate files for each partition, it
creates one file which simulates the whole disk. Then with "parted" util,
this disk is parted to desired partitions. The disk file is attached to
loop device with parted option. It allows to get partitions UUID and also
simulate the behaviour which is closed to the target device.

Partition and servicecontroller tests are rewritten with new testtools disk
approach.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>