[NUNCOTA-500] Fix returning success status on bad update image file
If update manager can't extract the image, it returns the success status.
But should return error. The problem is in state switch on upgrade init:
update handler switches to "finishState" and error is overridden there.
The fix is to terminate upgrade loop by calling "finishOperation" function
with appropriate error and return.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>