[tpmmodule] Add removing outdated certificate after adding new one
We have to not exceed max cert item in module. But it is too risky, first
remove old certificate and then add new one. If something bad happens in
the middle, we may loose all certificates. To avoid it, we add new cert
first and then remove olds by checking "notafter" cert field.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>