[updatehandler] Fix concurrent map access in prepareComponent function
prepareComponent is executed in parallel for all components in the current
update. As result handler.state.CurrentVendorVersions and
handler.componentStatuses may have concurrent access. Access to all common data
should be avoided or serialized in this function. The fix is to move handling of
handler.state.CurrentVendorVersions and handler.componentStatuses to
onPrepareState function which guarded with handler mutex.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>